feat: move tthemeId from common to data, rename files and directories, update the data
