+++ libs/common/src/index.ts
@@ -9,4 +9,3 @@ export * from './questionnaire/types/questionnaireData'
 export * from './questionnaire/types/types'
 export * from './questionnaire/objectiveChecker'
 export * from './stats/types'
-export * from './theme/types'
