+++ libs/data/src/index.ts
@@ -2,3 +2,4 @@ export * from './program/program'
 export * from './program/types'
 export * from './type/publicodesTypes'
 export * from './generated/program'
+export * from './theme/themes'
